requirements.txt
gensim==4.3.1
grpcio==1.56.0
grpcio-tools==1.56.0
joblib==1.3.1
numpy==1.25.1
protobuf==4.23.4
python-dotenv==1.0.0
scikit-learn==1.3.0
scipy==1.11.1
smart-open==6.3.0
threadpoolctl==3.2.0
